Fixed missing indexes
A few minor missing indexes.

For cleanliness and because the project is so early, I opted to make
this change inline.  That way there isn't an extra migration.  So a
'rake db:migrate:reset' is necessary.  If the project were any older, a
separate migration would be preferred, but this seems appropriate.