Fixed DB schema sync issue on Message
The 'from' column was removed from Message in a890eadb But the schema
and message model wasn't updated to reflect this.  Tests were still
passing because the schema had the column.  But a full db:migrate:reset
caused it to start failing.

This is the minimum necessary to make it work again.