Added webhook serializer
Webhook dynamically calculates its signature, however
that calculation is irrelevant for ensuring it is
serialized properly. So signature is stubbed out.

This required adding a stubbing/mocking framework.  Although
this wasn't officially discussed, I've run into this issue on
several occasion: needing to stub or mock where minitest/mock
was insufficient.

I chose flexmock because it's very small, lightweight,
has a namespace that doesn't conflict with any other library,
and served the purpose.

Should the community choose otherwise, it would be easy to replace
later with something else.