Conversation view implemented.
This is a massive commit, please do as I say, not as I do. It breaks a lot of existing functionality which I'm going to add back in ASAP. In particular:

* The homepage has changed.
* Some of he form styles are broken (but we're not actually using them anywhere)
* Keyboard shortcuts don't work
* Replying to messages doesn't work
* Search doesn't work

I think most of those are pretty trivial to add back in. However, this commit was out of time pressure, not because it's a great example of how we should work. Please bear that in mind.