Added API docs
This starts an outline of the API docs.

The formatting is Markdown as described by the latest Apiary Blueprint
format:  http://apiary.io/blueprint

This format allows for:
  * Easy to type/format
  * Machine readable/parsable docs
  * Viewable on Apiary.io

For local viewing, kept the ability to just look at it via /docs.  But
Apiary.io is much nicer and gives a UI for easy testing.