Switch from MiniTest to RSpec.
This is a big change, especially after a lot of previous discussion. I
think the reality of the situation is that:

* Minitest has lead to multiple conflicting styles
* Has poor integration with Rails
* Isn’t maintained

My central argument for picking Minitest was to stick with
lowest-common-denominator tools. I think at this point RSpec is known
by everybody and won’t impact the chances of somebody contributing.

I sat about trying to upgrade the codebase to Rails 4.1 and there are
multiple problems with trying to upgrade minitest-rails. I just decided
to bite the bullet and convert.

As to sticking to a standard, I think we should aim to abide by the
guide at http://betterspecs.org. They seem reasonable to me.