Added jazz_hands suite of tools and awesome_print
jazz_hands is a curated suite of tools for Rails. It adds Hirb, Awesome
Print, Pry and a bundle of Pry extensions that are super useful.

awesome_print is explicitly added into the production environment
because it provides a logging extesion that is incredibly useful for
reading hard-to-parse complex structures in logs.
`log.ap(Message.new, :debug)` for an example. Hirb is configured to be
automatically enabled in development and testing.

I also configure Pry as the default console shell for all environents
because it is just so darned powerful.